+++ internal/rag/embedder/testmain_test.go
@@ -4,10 +4,10 @@ import (
 	"os"
 	"testing"
 
-	chassis "github.com/ai8future/chassis-go"
+	chassis "github.com/ai8future/chassis-go/v5"
 )
 
 func TestMain(m *testing.M) {
-	chassis.RequireMajor(4)
+	chassis.RequireMajor(5)
 	os.Exit(m.Run())
 }